Split stateful tests into validat12.cpp
Move the LMS, HSS, and FileStateStore tests out of validat11.cpp and wire the new file into the test builds.
