+++ cryptest.vcxproj
@@ -228,6 +228,7 @@
     <ClCompile Include="src\test\validat10.cpp" />
     <ClCompile Include="src\test\validat_cve_2024_28285.cpp" />
     <ClCompile Include="src\test\validat11.cpp" />
+    <ClCompile Include="src\test\validat12.cpp" />
   </ItemGroup>
   <!-- Header Files -->
   <ItemGroup>
