+++ cryptest.vcxproj.filters
@@ -527,6 +527,9 @@
     <ClCompile Include="src\test\validat11.cpp">
       <Filter>Source Code</Filter>
     </ClCompile>
+    <ClCompile Include="src\test\validat12.cpp">
+      <Filter>Source Code</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="bench.h">
